Move PowerBiService to services Add PowerBiService as singleton
